create PayloadNormalizer.php, TaxonResponseHandler.php, TaxonViewDataService.php to simplify TaxonomyController further and then improve robustness of e2e tests
