+++ e2e/taxonomy-create-edit-submit.spec.cjs
@@ -1,35 +1,13 @@
 // @ts-check
 const { test, expect } = require('@playwright/test');
-
-const TEST_EMAIL = process.env.TEST_USER_EMAIL || 'mark.fisher@ku.edu';
-const TEST_PASSWORD = process.env.TEST_USER_PASSWORD || 'tomcat123';
-let cachedCookies = null;
-
-async function login(page) {
-    if (cachedCookies) {
-        await page.context().addCookies(cachedCookies);
-        await page.goto('/');
-        await page.waitForLoadState('networkidle');
-        return;
-    }
-
-    await page.goto('/login');
-    await page.locator('#email').fill(TEST_EMAIL);
-    await page.locator('#password').fill(TEST_PASSWORD);
-    await Promise.all([
-        page.waitForURL((url) => !url.pathname.includes('login'), { timeout: 20_000 }),
-        page.locator('form').first().evaluate((form) => form.submit()),
-    ]);
-    await page.waitForLoadState('networkidle');
-    cachedCookies = await page.context().cookies();
-}
+const { login } = require('./helpers/auth.cjs');
 
 async function createTaxon(page) {
     const suffix = Date.now();
     const randomToken = Math.random().toString(36).slice(2, 8);
     const speciesEpithet = `copilot${suffix}-${randomToken}`;
 
-    await page.goto('/taxon/create');
+    await page.goto('/taxon/create', { waitUntil: 'domcontentloaded' });
     await expect(page.locator('#taxon-form')).toBeVisible({ timeout: 15_000 });
 
     const quickParserValue = `Drosera ${speciesEpithet}`;
@@ -38,31 +16,19 @@ async function createTaxon(page) {
 
     await expect(page.locator('#unitname1')).toHaveValue('Drosera', { timeout: 10_000 });
     await expect(page.locator('#unitname2')).toHaveValue(speciesEpithet, { timeout: 10_000 });
-    await page.locator('#parentname').fill('Drosera');
-    await page.waitForResponse((resp) => resp.url().includes('/api/taxa/search'), { timeout: 5000 });
-
-    const parentResults = page.locator('#search-results-parentname > *');
-    const parentResultCount = await parentResults.count();
-    if (parentResultCount > 0) {
-        const firstParentResult = parentResults.first();
-        const selectedParentTid = await firstParentResult.getAttribute('id');
-        expect(selectedParentTid).toBeTruthy();
-        await firstParentResult.click();
-    } else {
-        await page.locator('#parentname').fill('Organism');
-        await page.locator('#tid-parentname').evaluate((el) => {
-            el.value = '1';
-            el.dispatchEvent(new Event('input', { bubbles: true }));
-            el.dispatchEvent(new Event('change', { bubbles: true }));
-        });
-    }
+    await page.locator('#parentname').fill('Organism');
+    await page.locator('#tid-parentname').evaluate((el) => {
+        el.value = '1';
+        el.dispatchEvent(new Event('input', { bubbles: true }));
+        el.dispatchEvent(new Event('change', { bubbles: true }));
+    });
 
     await page.locator('#parentname').blur();
     await expect(page.locator('#tid-parentname')).toHaveValue(/\d+/);
     await expect(page.locator('#submitButton')).toBeEnabled({ timeout: 10_000 });
 
     await Promise.all([
-        page.waitForURL(/\/taxon\/\d+$/, { timeout: 15_000 }),
+        page.waitForURL(/\/taxon\/\d+$/, { timeout: 30_000, waitUntil: 'domcontentloaded' }),
         page.locator('#taxon-form').evaluate((form) => form.submit()),
     ]);
